state-res: Move state resolution code to its own module
